Revert changes caused by rebase
