+++ packages/block-editor/src/hooks/use-bindings-attributes.js
@@ -11,11 +11,6 @@ import { addFilter } from '@wordpress/hooks';
  * Internal dependencies
  */
 import { unlock } from '../lock-unlock';
-import {
-	canBindBlock,
-	canBindAttribute,
-	getBoundAttributesValues,
-} from '../utils/bindings';
 
 /** @typedef {import('@wordpress/compose').WPHigherOrderComponent} WPHigherOrderComponent */
 /** @typedef {import('@wordpress/blocks').WPBlockSettings} WPBlockSettings */
@@ -27,26 +22,90 @@ import {
  * @return {WPHigherOrderComponent} Higher-order component.
  */
 
+const BLOCK_BINDINGS_ALLOWED_BLOCKS = {
+	'core/paragraph': [ 'content' ],
+	'core/heading': [ 'content' ],
+	'core/image': [ 'id', 'url', 'title', 'alt' ],
+	'core/button': [ 'url', 'text', 'linkTarget', 'rel' ],
+};
+
+/**
+ * Based on the given block name,
+ * check if it is possible to bind the block.
+ *
+ * @param {string} blockName - The block name.
+ * @return {boolean} Whether it is possible to bind the block to sources.
+ */
+export function canBindBlock( blockName ) {
+	return blockName in BLOCK_BINDINGS_ALLOWED_BLOCKS;
+}
+
+/**
+ * Based on the given block name and attribute name,
+ * check if it is possible to bind the block attribute.
+ *
+ * @param {string} blockName     - The block name.
+ * @param {string} attributeName - The attribute name.
+ * @return {boolean} Whether it is possible to bind the block attribute.
+ */
+export function canBindAttribute( blockName, attributeName ) {
+	return (
+		canBindBlock( blockName ) &&
+		BLOCK_BINDINGS_ALLOWED_BLOCKS[ blockName ].includes( attributeName )
+	);
+}
+
 export const withBlockBindingSupport = createHigherOrderComponent(
 	( BlockEdit ) => ( props ) => {
 		const registry = useRegistry();
-		const {
-			attributes: blockAttributes,
-			name,
-			clientId,
-			context,
-			setAttributes,
-		} = props;
-		const bindings = blockAttributes?.metadata?.bindings;
-
-		// It seems if I don't wrap this in a useSelect, the reset in pattern overrides doesn't work as expected.
+		const sources = useSelect( ( select ) =>
+			unlock( select( blocksStore ) ).getAllBlockBindingsSources()
+		);
+		const bindings = props.attributes.metadata?.bindings;
+		const { name, clientId, context } = props;
 		const boundAttributes = useSelect( () => {
-			return getBoundAttributesValues( clientId, context, registry );
-		}, [ clientId, context, registry ] );
+			if ( ! bindings ) {
+				return;
+			}
 
-		const sources = unlock(
-			registry.select( blocksStore )
-		).getAllBlockBindingsSources();
+			const attributes = {};
+
+			for ( const [ attributeName, boundAttribute ] of Object.entries(
+				bindings
+			) ) {
+				const source = sources[ boundAttribute.source ];
+				if (
+					! source?.getValue ||
+					! canBindAttribute( name, attributeName )
+				) {
+					continue;
+				}
+
+				const args = {
+					registry,
+					context,
+					clientId,
+					attributeName,
+					args: boundAttribute.args,
+				};
+
+				attributes[ attributeName ] = source.getValue( args );
+
+				if ( attributes[ attributeName ] === undefined ) {
+					if ( attributeName === 'url' ) {
+						attributes[ attributeName ] = null;
+					} else {
+						attributes[ attributeName ] =
+							source.getPlaceholder?.( args );
+					}
+				}
+			}
+
+			return attributes;
+		}, [ bindings, name, clientId, context, registry, sources ] );
+
+		const { setAttributes } = props;
+
 		const _setAttributes = useCallback(
 			( nextAttributes ) => {
 				registry.batch( () => {
@@ -131,7 +190,7 @@ export const withBlockBindingSupport = createHigherOrderComponent(
 			<>
 				<BlockEdit
 					{ ...props }
-					attributes={ { ...blockAttributes, ...boundAttributes } }
+					attributes={ { ...props.attributes, ...boundAttributes } }
 					setAttributes={ _setAttributes }
 				/>
 			</>
@@ -161,6 +220,6 @@ function shimAttributeSource( settings, name ) {
 
 addFilter(
 	'blocks.registerBlockType',
-	'core/editor/use-bindings-attributes',
+	'core/editor/custom-sources-backwards-compatibility/shim-attribute-source',
 	shimAttributeSource
 );
